ai-accounting-agent/ai-accountant-mcp/pyproject.toml
[project]
name = "ai-accountant-mcp"
version = "0.1.0"
description = "A MCP for server using Tripletex API v2"
readme = "README.md"
authors = [
    { name = "valiantlynx", email = "valiantlynxz@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.0.2",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"
